修改代码和mapper文件

+++ src/main/java/com/zes/squad/gmh/web/entity/po/StaffPo.java
@@ -30,7 +30,7 @@ public class StaffPo extends Po {
     /**
      * 等级
      */
-    private Long staffLevel;
+    private Integer           staffLevel;
     /**
      * 负责人姓名
      */
